+++ trace-processor/common/src/main/java/eu/csgroup/coprs/monitoring/common/datamodel/entities/Product.java
@@ -68,8 +68,6 @@ public class Product implements DefaultEntity, Serializable {
 
     private Instant evictionDate;
 
-    private boolean late;
-
     @Override
     public Product copy() {
         return this.toBuilder().build();